+++ yaml_ld/document_loaders/local_file.py
@@ -10,19 +10,13 @@ from yaml.parser import ParserError
 from yaml.scanner import ScannerError
 
 from yaml_ld.document_loaders.base import DocumentLoader, PyLDResponse
+from yaml_ld.document_parsers.html_parser import HTMLDocumentParser
 from yaml_ld.document_parsers.yaml_parser import YAMLDocumentParser
 from yaml_ld.load_html import load_html
 from yaml_ld.loader import YAMLLDLoader
 
 
 class LocalFileDocumentLoader(DocumentLoader):
-    def _parse_script_content(self, content: str):
-        return list(
-            yaml.load_all(
-                content,
-                Loader=YAMLLDLoader,
-            ),
-        )
 
     def __call__(self, source: str | Path, options: dict[str, Any]) -> PyLDResponse:
         from yaml_ld.errors import DocumentIsScalar, LoadingDocumentFailed
@@ -32,7 +26,7 @@ class LocalFileDocumentLoader(DocumentLoader):
         if path.suffix in {'.yaml', '.yml', '.yamlld', '.json', '.jsonld'}:
             try:
                 with path.open() as f:
-                    yaml_document = YAMLDocumentParser()(f, options)
+                    yaml_document = YAMLDocumentParser()(f, source, options)
 
                     return {
                         'document': yaml_document,
@@ -46,17 +40,7 @@ class LocalFileDocumentLoader(DocumentLoader):
 
         if path.suffix in {'.html', '.xhtml'}:
             with path.open() as f:
-                loaded_html = load_html(
-                    input=f.read(),
-                    url=source,
-                    profile=None,
-                    options=options,
-                    content_type='application/ld+yaml',
-                    parse_script_content=self._parse_script_content,
-                )
-
-                if isinstance(loaded_html, str):
-                    raise DocumentIsScalar(loaded_html)
+                loaded_html = HTMLDocumentParser()(f, source, options)
 
             return {
                 'document': loaded_html,
