+++ yaml_ld/document_parsers/base.py
@@ -7,5 +7,5 @@ from yaml_ld.models import Document
 
 class BaseDocumentParser:
     @abstractmethod
-    def __call__(self, data: io.TextIOBase, options: dict[str, Any]) -> Document:
+    def __call__(self, data: io.TextIOBase, source: str, options: dict[str, Any]) -> Document:
         raise NotImplementedError()
